chore(cli): move cmd to internal/cli
