+++ .gitignore
@@ -2,7 +2,7 @@
 transit
 
 # Local CLI config file (useful for symlinking to config dir while developing)
-config.*
+/config.*
 
 # goreleaser compiled binaries go here
 dist/
