+++ .goreleaser.yaml
@@ -31,7 +31,7 @@ builds:
       - -trimpath
     ldflags:
       - -s -w
-      - -X github.com/ismailshak/transit/cmd.version={{.Version}}
+      - -X github.com/ismailshak/transit/internal/cli.version={{.Version}}
 
 archives:
   - name_template: >-
